+++ src/common/localization/en-us.ts
@@ -9,7 +9,7 @@ import { IAppStrings } from "../strings";
  * App Strings for English language
  */
 export const english: IAppStrings = {
-    appName: "OCR Form Labeling Tool",
+    appName: "Form OCR Testing Tool",
     common: {
         displayName: "Display name",
         description: "Description",
@@ -70,6 +70,7 @@ export const english: IAppStrings = {
             key: {
                 title: "Key",
             },
+            duplicateNameErrorMessage: "Token name must be unique for all tokens",
         },
         securityTokens: {
             title: "Security Tokens",
@@ -164,12 +165,21 @@ export const english: IAppStrings = {
             titleAria: "Compose Model View",
             addComposeModelName: "Add compose model name...",
             NotEnoughModels: " Should have at least more than one selected model to compose a new model",
+            modelsCannotBeIncluded: "Warning: These models will not be included in composed model!",
+            modelCannotBeIncluded: "Warning: This model will not be included in composed model!"
         },
         commandBar: {
             ariaLabel: "Please use command bar to compose models",
             composeAria: "Compose Model",
             refreshAria: "Refresh the list",
             filter: "Filter by name...",
+            filterAria: "Filter by name input area",
+        },
+        modelsList: {
+            headerAria: "List of models header",
+            checkButtonAria: "Select model check button",
+            checkAllButtonAria: "Select all models check button",
+
         }
     },
     predict: {
@@ -246,6 +256,7 @@ export const english: IAppStrings = {
         instructions: "Please select a connection to edit",
         new: "New Connection",
         save: "Save Connection",
+        genericInvalid: "\"${project.sourceConnection.name}\" is an invalid connection. Please check it in the Connections page",
         messages: {
             saveSuccess: "Successfully saved ${connection.name}",
             deleteSuccess: "Successfully deleted ${connection.name}",
@@ -276,6 +287,7 @@ export const english: IAppStrings = {
                     title: "Create Container",
                     description: "Creates the blob container if it does not already exist",
                 },
+                invalidSASMessage: "\"${project.sourceConnection.name}\" has no storage account. Please check it's SAS token in the Connections page",
             },
             bing: {
                 title: "Bing Image Search",
@@ -292,9 +304,11 @@ export const english: IAppStrings = {
             },
             local: {
                 title: "Local file system",
-                folderPath: "Browse",
+                folderPath: "Folder",
+                browse: "Browse",
                 selectFolder: "Select folder",
                 chooseFolder: "Choose folder",
+                invalidFolderMessage: "\"${project.sourceConnection.name}\" has an invalid folder. Please check it's selected folder in the Connections page",
             },
         },
     },