+++ src/common/mockFactory.ts
@@ -147,6 +147,7 @@ export default class MockFactory {
             apiUriBase: "localhost",
             folderPath: "",
             trainRecord: null,
+            recentModelRecords: [],
             predictModelId: "",
         };
     }
@@ -311,6 +312,7 @@ export default class MockFactory {
             readBinary: jest.fn(),
             deleteFile: jest.fn(),
             writeText: jest.fn(),
+            isValidProjectConnection: jest.fn(),
             writeBinary: jest.fn(),
             listFiles: jest.fn(() => Promise.resolve(MockFactory.createFileList())),
             listContainers: jest.fn(),