+++ src/common/strings.ts
@@ -70,6 +70,7 @@ export interface IAppStrings {
             key: {
                 title: string;
             },
+            duplicateNameErrorMessage: string,
         },
         securityTokens: {
             title: string;
@@ -164,13 +165,21 @@ export interface IAppStrings {
             titleAria: string;
             addComposeModelName: string;
             NotEnoughModels: string;
+            modelsCannotBeIncluded: string;
+            modelCannotBeIncluded: string;
         }
         commandBar: {
             ariaLabel: string;
             composeAria: string;
             refreshAria: string;
             filter: string;
-        }
+            filterAria: string;
+        },
+        modelsList: {
+            headerAria: string,
+            checkButtonAria: string,
+            checkAllButtonAria: string,
+        },
     }
     predict: {
         title: string;
@@ -246,6 +255,7 @@ export interface IAppStrings {
         instructions: string;
         new: string;
         save: string;
+        genericInvalid: string;
         messages: {
             saveSuccess: string;
             deleteSuccess: string;
@@ -272,7 +282,8 @@ export interface IAppStrings {
                 createContainer: {
                     title: string,
                     description: string,
-                }
+                },
+                invalidSASMessage: string;
             },
             bing: {
                 title: string;
@@ -290,8 +301,10 @@ export interface IAppStrings {
             local: {
                 title: string;
                 folderPath: string;
+                browse: string;
                 selectFolder: string;
                 chooseFolder: string;
+                invalidFolderMessage: string;
             },
         }
     };