+++ src/providers/storage/localFileSystemProxy.ts
@@ -49,6 +49,10 @@ export class LocalFileSystemProxy implements IStorageProvider, IAssetProvider {
         return IpcRendererProxy.send(`${PROXY_NAME}:readText`, [filePath]);
     }
 
+    public isValidProjectConnection(): Promise<boolean> {
+        return IpcRendererProxy.send(`${PROXY_NAME}:isValidProjectConnection`, [this.options.folderPath]);
+    }
+
     public getFileType(fileName: string): Promise<any> {
         const filePath = [this.options.folderPath, fileName].join("/");
         return IpcRendererProxy.send(`${PROXY_NAME}:getFileType`, [filePath]);