+++ src/react/components/pages/predict/predictResult.tsx
@@ -6,10 +6,18 @@ import { ITag } from "../../../../models/applicationState";
 import "./predictResult.scss";
 import { getPrimaryGreenTheme } from "../../../../common/themes";
 import { PrimaryButton } from "@fluentui/react";
+import PredictModelInfo from './predictModelInfo';
+
+export interface IAnalyzeModelInfo {
+    docType: string,
+    modelId: string,
+    docTypeConfidence: number,
+}
 
 export interface IPredictResultProps {
     predictions: { [key: string]: any };
     analyzeResult: {};
+    analyzeModelInfo: IAnalyzeModelInfo;
     page: number;
     tags: ITag[];
     downloadResultLabel: string;
@@ -22,7 +30,7 @@ export interface IPredictResultState { }
 
 export default class PredictResult extends React.Component<IPredictResultProps, IPredictResultState> {
     public render() {
-        const { tags, predictions } = this.props;
+        const { tags, predictions, analyzeModelInfo } = this.props;
         const tagsDisplayOrder = tags.map((tag) => tag.name);
         for (const name of Object.keys(predictions)) {
             const prediction = predictions[name];
@@ -47,6 +55,7 @@ export default class PredictResult extends React.Component<IPredictResultProps, 
                         onClick={this.triggerDownload}
                     />
                 </div>
+                <PredictModelInfo modelInfo={analyzeModelInfo} />
                 <div className="prediction-field-header">
                     <h6 className="prediction-field-header-field"> Page # / Field name / Value</h6>
                     <h6 className="prediction-field-header-confidence"> Confidence %</h6>